repr: don't depend on Avro
mz-repr is a foundational crate that's meant to be light on
dependencies. It should not have a dependency on mz-avro, as that's not
a core part of the system, but rather just something you use with
`FORMAT AVRO`.

The dependency was introduced in 41be1fa7, in order to facilitate
Avro decoding for CDCv2.

This commit moves the Avro decoding implementation for the Timestamp
type out of mz-repr and into mz-interchange, the one crate that actually
needs it.

Fix #16521.
