+++ src/repr/Cargo.toml
@@ -27,7 +27,6 @@ fast-float = "0.2.0"
 hex = "0.4.3"
 itertools = "0.10.5"
 once_cell = "1.16.0"
-mz-avro = { path = "../avro" }
 mz-lowertest = { path = "../lowertest" }
 mz-ore = { path = "../ore", features = ["bytes", "smallvec", "stack", "test"] }
 mz-persist-types = { path = "../persist-types" }
